pyproject.toml
[tool.poetry]
name = "live-speed-limit-sensor"
version = "0.1.0"
description = ""
authors = ["amoghstayade <tayade.amogh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
pre-commit = "^2.20.0"
python-dotenv = "^0.20.0"
serial = "^0.0.97"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
